+++ test/sweatkit/test_util.clj
@@ -1,5 +1,0 @@
-(ns sweatkit.test-util
-  (:require [clojure.java.io :as io]))
-
-(defn read-file [path]
-  (io/file path))